Enabled providers and tests for identities page

+++ app/Http/Controllers/Profile/UserIdentitiesController.php
@@ -3,6 +3,7 @@
 namespace KBox\Http\Controllers\Profile;
 
 use Illuminate\Http\Request;
+use KBox\Facades\Identity as FacadesIdentity;
 use KBox\Http\Controllers\Controller;
 use KBox\Identity;
 
@@ -21,12 +22,18 @@ class UserIdentitiesController extends Controller
      */
     public function index(Request $request)
     {
+        $available = FacadesIdentity::enabledProviders();
+
+        abort_if(empty($available), 404);
+
         $identities = $request->user()->identities;
 
         return view('profile.identities', [
             'pagetitle' => trans('profile.identities'),
             'breadcrumb_current' => trans('profile.identities'),
-            'identities' => $identities
+            'identities' => $identities,
+            'enabled' => $identities->pluck('provider'),
+            'availableProviders' => $available,
         ]);
     }
 
@@ -38,6 +45,10 @@ class UserIdentitiesController extends Controller
      */
     public function destroy(Request $request, Identity $identity)
     {
+        $available = FacadesIdentity::enabledProviders();
+
+        abort_if(empty($available), 404);
+        
         $this->authorize('delete', $identity);
 
         $identity->delete();
